# coursework/backend/src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: Krusty Crab CRM API
  description: |
    API для системы управления рестораном "Красти Крабс"
    
    Система предназначена для автоматизации бизнес-процессов ресторана:
    - Управление заказами и клиентами
    - Контроль меню и ингредиентов
    - Управление сотрудниками и рабочими сменами
    - Финансовый учет и отчётность
    - Программы лояльности
    - Мониторинг качества обслуживания
  version: 1.0.0
  contact:
    name: Krusty Crab Support
servers:
  - url: http://localhost:13228/api
    description: Local development server
...
paths:
...
  /auth/login/client:
    post:
      tags:
        - auth
      summary: Вход клиента
      description: Аутентификация клиента по email и паролю
      operationId: loginClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT токен доступа
                  userType:
                    type: string
                    enum: [CLIENT]
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Клиент не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login/employee:
    post:
      tags:
        - auth
      summary: Вход сотрудника
      description: Аутентификация сотрудника по логину и паролю
      operationId: loginEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
              properties:
                login:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT токен доступа
                  userType:
                    type: string
                    enum: [EMPLOYEE]
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Сотрудник не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags:
        - auth
      summary: Получить информацию о текущем пользователе
      description: Получение информации о текущем аутентифицированном пользователе
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация о пользователе
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  username:
                    type: string
                  userType:
                    type: string
                    enum: [CLIENT, EMPLOYEE]
                  role:
                    type: string
                    nullable: true
                    description: Роль сотрудника (только для EMPLOYEE)
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== CLIENTS ==========
  # ========== CLIENTS ==========
...
  /analytics/sales:
    get:
      tags:
        - analytics
      summary: Получить сводку по продажам
      description: Получение сводки по продажам за период (FR-014, FR-020, для управляющего)
      operationId: getSalesSummary
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Сводка по продажам
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesSummary'
  /analytics/top-menu-items:
    get:
      tags:
        - analytics
      summary: Получить топ блюд
      description: Получение топ блюд по количеству и выручке (FR-012, FR-020, для управляющего)
      operationId: getTopMenuItems
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Количество записей
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Топ блюд
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopMenuItem'
  /analytics/sales/by-employee:
    get:
      tags:
        - analytics
      summary: Продажи по сотрудникам
      description: Разрез продаж по сотрудникам (кассир/управляющий) за период (UC-4)
      operationId: getSalesByEmployee
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Продажи по сотрудникам
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesByEmployeeItem'
  /analytics/sales/by-time-of-day:
    get:
      tags:
        - analytics
      summary: Продажи по времени суток
      description: Разрез продаж по времени суток (morning/day/evening/night) за период (UC-4)
      operationId: getSalesByTimeOfDay
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Продажи по времени суток
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesByTimeOfDayItem'
  /analytics/report-views:
    get:
      tags:
        - analytics
      summary: История просмотров отчётов
      description: Получение истории просмотров отчётов (UC-4 шаг 4)
      operationId: getReportViews
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Максимум записей
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 500
        - name: offset
          in: query
          description: Смещение (пагинация)
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: История просмотров отчётов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportView'
  /analytics/financial:
    get:
      tags:
        - analytics
      summary: Финансовая сводка
      description: Финансовый отчёт за период (выручка + расходы) (UC-13)
      operationId: getFinancialSummary
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          description: Начало периода
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Конец периода
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Финансовая сводка
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialSummary'

  # ========== FINANCE ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен доступа. Получите токен через /auth/login/client или /auth/login/employee
...
  schemas:
...
    ReportView:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        report:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        viewedAt:
          type: string
          format: date-time
    FinancialSummary:
      type: object
      properties:
        hasData:
          type: boolean
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        revenue:
          type: number
          format: decimal
        ingredientExpenses:
          type: number
          format: decimal
        salaryExpenses:
          type: number
          format: decimal
        profit:
          type: number
          format: decimal
...
    SalesSummary:
      type: object
      properties:
        hasData:
          type: boolean
        fromTs:
          type: string
          format: date-time
        toTs:
          type: string
          format: date-time
        ordersCnt:
          type: integer
        revenue:
          type: number
          format: decimal
        avgTicket:
          type: number
          format: decimal
    SalesByEmployeeItem:
      type: object
      properties:
        employeeId:
          type: integer
        employeeName:
          type: string
        ordersCnt:
          type: integer
        revenue:
          type: number
          format: decimal
        avgTicket:
          type: number
          format: decimal
    SalesByTimeOfDayItem:
      type: object
      properties:
        bucket:
          type: string
          description: Время суток (morning/day/evening/night)
        ordersCnt:
          type: integer
        revenue:
          type: number
          format: decimal
        avgTicket:
          type: number
          format: decimal
    TopMenuItem:
      type: object
      properties:
        menuItemId:
          type: integer
        name:
          type: string
        quantity:
          type: integer
        revenue:
          type: number
          format: decimal
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string